organize background components
