+++ app/src/main/java/edu/rpi/shuttletracker/app/ShuttleTrackerApplication.kt
@@ -4,9 +4,9 @@ import android.app.Application
 import androidx.hilt.work.HiltWorkerFactory
 import androidx.work.Configuration
 import dagger.hilt.android.HiltAndroidApp
-import edu.rpi.shuttletracker.util.notifications.Notifications
-import edu.rpi.shuttletracker.util.services.FirebaseService
-import edu.rpi.shuttletracker.util.workers.AnnouncementWorker
+import edu.rpi.shuttletracker.background.notification.Notifications
+import edu.rpi.shuttletracker.background.service.FirebaseService
+import edu.rpi.shuttletracker.background.worker.AnnouncementWorker
 import javax.inject.Inject
 
 @HiltAndroidApp
